build cleanup also trims the history index correctly

+++ src/Modules/BuildCleanup/Model/BuildCleanupLinuxUnix.php
@@ -29,6 +29,12 @@ class BuildCleanupLinuxUnix extends Base {
                 "optional" => true,
                 "name" => "Enable Removal of Old Builds?"
             ),
+            "trim_history_index" =>
+            array(
+                "type" => "boolean",
+                "optional" => true,
+                "name" => "Also trim History Index? May restrict monitoring data"
+            ),
             "no_to_keep" =>
             array(
                 "type" => "text",
@@ -126,24 +132,36 @@ class BuildCleanupLinuxUnix extends Base {
         $keeps = array_slice($arr, 0, $nk);
         $drops = array_diff($historyFiles, $keeps);
         foreach ($drops as $dropfile) {
-            $this->lm->log ("Removing Steps History file {$dropfile}", $this->getModuleName() ) ;
-            $rmCommand = 'rm -f '.PIPEDIR.DS.$this->params["item"].DS.'stepsHistory'.DS.$dropfile ;
-            self::executeAndOutput($rmCommand) ; }
+            if (!in_array($dropfile, array(".", ".."))) {
+                $this->lm->log ("Removing Steps History file {$dropfile}", $this->getModuleName() ) ;
+                $rmCommand = 'rm -f '.PIPEDIR.DS.$this->params["item"].DS.'stepsHistory'.DS.$dropfile ;
+                self::executeAndOutput($rmCommand) ; } }
+        // @todo is it really tho
         $result = true ;
         return $result ;
     }
 
     private function trimHistoryIndex() {
-        // @todo this doesnt do anythin
-//        $historyFiles = scandir(PIPEDIR.DS.$this->params["item"].DS.'history') ;
-//        asort($historyFiles);
-//        $arr = array_reverse($historyFiles);
-//        $keeps = array_slice($arr, 0, 3);
-//        $drops = array_diff($historyFiles, $keeps);
-//        foreach ($drops as $dropfile) {
-//            $this->lm->log ("Removing History file {$dropfile}", $this->getModuleName() ) ;
-//            $rmCommand = 'rm '.$dropfile ;
-//            self::executeAndOutput($rmCommand) ; }
+        // @todo this doesn't do anything
+        $mn = $this->getModuleName() ;
+        if (isset($this->params["build-settings"][$mn]["trim_history_index"]) &&
+            $this->params["build-settings"][$mn]["trim_history_index"] == "on") {
+            $this->lm->log("Trimming History Index is enabled, trimming", $this->getModuleName() ) ;
+            $historyFile = file_get_contents(PIPEDIR.DS.$this->params["item"].DS.'historyIndex') ;
+            $arr = json_decode($historyFile, true);
+            if (isset($this->params["build-settings"][$mn]["no_to_keep"]) &&
+                is_int($this->params["build-settings"][$mn]["no_to_keep"])) {
+                $nk = $this->params["build-settings"][$mn]["no_to_keep"]; }
+            else {
+                $this->lm->log ("No amount of builds to keep has been set, defaulting to three", $this->getModuleName() ) ;
+                $nk = 3; }
+            $keeps = array_slice($arr, 0, $nk);
+            $countDrops = count(array_diff($arr, $keeps)) ;
+            $this->lm->log ("Removing {$countDrops} records from the history index", $this->getModuleName() ) ;
+            $arr = json_encode($keeps);
+            file_put_contents(PIPEDIR.DS.$this->params["item"].DS.'historyIndex', $arr) ; }
+        else {
+            $this->lm->log ("Trimming History Index is not enabled", $this->getModuleName() ) ; }
         $result = true ;
         return $result ;
     }
